Ajustado Delete para retornar 422 em caso de erro

+++ backend/app/Http/Controllers/api/ProductController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\api;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Product;
+use Illuminate\Support\Facades\DB;
+use \Illuminate\Http\Response;
 
 class ProductController extends Controller
 {
@@ -32,7 +34,23 @@ class ProductController extends Controller
 
     public function destroy($id)
     {
-        $product = Product::findOrFail($id);
-        $product->delete();
+        $valida = DB::table('products')
+        ->select(DB::raw('count(orders.id) as qtdpedidos'))
+        ->join('orders', 'orders.id_product', '=', 'products.id')
+        ->where('products.id', '=', $id)
+        ->get();
+
+        foreach($valida as $temp){
+            $var = $temp->qtdpedidos;
+        }
+
+        if($var == 0){
+            $product = Product::findOrFail($id);
+            $product->delete();
+            return Product::all();
+        }else{
+            return  response()->json(['success' => false, 'detail' => (string) 'Produto com pedidos vinculados'], 422);
+        }
+
     }
 }
